Refacor Project, Configure Package
