+++ .gitignore
@@ -1,0 +1,4 @@
+.vscode
+dist
+__pycache__
+*.egg-info
