+++ pyproject.toml
@@ -1,0 +1,30 @@
+[build-system]
+requires = ["setuptools"]
+build-backend = "setuptools.build_meta"
+
+[project]
+name = "tucciafo_diffusion2d"
+authors = [
+    {name = "Fabio Tucciarone", email = "177167@stud.uni-stuttgart.de"},
+]
+description = "This code solves the diffusion equation over a two dimensional square domain which is at a certain temperature, and a circular disc at its center which is at a higher temperature. The diffusion equation is solved using the finite-difference method. The thermal diffusivity and initial conditions of the system can be changed by the user. The code produces four plots at various timepoints of the simulation. The diffusion process can be clearly observed in these plots."
+readme = "README.md"
+requires-python = ">=3.6"
+keywords = ["diffusion", "visualization"]
+license-files = ["LICENSE"]
+classifiers = [
+    "Development Status :: 2 - Pre-Alpha",
+    "Intended Audience :: Education",
+    "Programming Language :: Python :: 3",
+]
+dependencies = [
+    "matplotlib", 
+    "numpy",
+]
+version = "0.0.2"
+
+[project.urls]
+Repository = "https://github.com/FabioTucciarone/diffusion2d.git"
+
+[project.scripts]
+solve-diffusion2d = "tucciafo_diffusion2d.diffusion2d:solve"
