+++ tucciafo_diffusion2d/__init__.py
@@ -1,0 +1,1 @@
+import tucciafo_diffusion2d
