+++ tucciafo_diffusion2d/output.py
@@ -1,0 +1,23 @@
+
+from matplotlib.figure import Figure
+from matplotlib.image import AxesImage
+import matplotlib.pyplot as plt
+import numpy as np
+
+
+def create_plot(fig: Figure, fig_counter: int, u: np.ndarray, T_cold: float, T_hot: float, n: int, dt: float):
+    ax = fig.add_subplot(220 + fig_counter)
+    im = ax.imshow(u.copy(), cmap=plt.get_cmap('hot'), vmin=T_cold, vmax=T_hot)  # image for color bar axes
+    ax.set_axis_off()
+    ax.set_title('{:.1f} ms'.format(n * dt * 1000))
+
+    return im
+
+
+def output_plots(fig: Figure, im: AxesImage | None):
+    if im is None:
+        return
+    fig.subplots_adjust(right=0.85)
+    cbar_ax = fig.add_axes((0.9, 0.15, 0.03, 0.7))
+    cbar_ax.set_xlabel('$T$ / K', labelpad=20)
+    fig.colorbar(im, cax=cbar_ax)
